Implement initial meta components
Implement a prototype ComponentLabel component under CanvasComponentButton to allow for CCButton to have relative positioned text in it. Meta components such as ComponentLabel do not get added to the global CanvasComponentCollection, but are instead found under their respective parent components such as CCButton in this case. Meta component Update and Draw are called but their parent components.